refactored codebase
