+++ .env.dev
@@ -1,2 +1,2 @@
-APOLLO_SERVER=https://172.76.10.161:4000/graphql
-APOLLO_SOCKET=wss://172.76.10.161:4000/graphql
+APOLLO_SERVER=https://172.76.10.161:4003/graphql
+APOLLO_SOCKET=wss://172.76.10.161:4003/graphql
