+++ src/views/Chat/components/ChatView/components/index.js
@@ -1,1 +1,6 @@
-export { default as Header } from './components/Header'
+export { default as EditorChat } from './EditorChat'
+export { default as GroupMessage } from './GroupMessage'
+export { default as Header } from './Header'
+export { default as MessageCard } from './MessageCard'
+export { default as PickerEmoji } from './PickerEmoji'
+export { default as ViewMessage } from './ViewMessage'
